Everyone Copy Cells in Here
I did some data cleaning, renaming of columns, creation of new ones that I think could be useful for our analysis, everyone should copy these cells into their main file and re-save the dataset from Friday after running these cells. It should have the cleaned data set be written overtop the old one, as long as the names are the same. This was needed as many of the numerical based categories were actually in string format, and I also wrote a function that defines the opponent for each player so we can calculate defensive statistics later on.